jetpack navigation 오류 수정
